requirements.txt
pymongo==2.8
requests==2.5.1
requests-oauthlib==0.4.2
setuptools==1.1
TwitterSearch==1.0.0
DateTime==4.0.1
zope.interfaces==4.1.2
oauthlib==0.7.2
python-dateutil==2.4.1
simplejson==3.6.5
git+git://github.com/AlchemyAPI/alchemyapi_python.git